+++ leaf.go
@@ -176,14 +176,14 @@ func (l *leaf) with(v interface{}, f func(a, b interface{}) interface{}, depth i
 	last := result
 	noffset, offset := h0.hash(), h.hash()
 	for noffset == offset {
-		last.mask = 1 << offset
+		last.mask = 1 << uint(offset)
 		newLast := &node{}
 		last.children[offset] = newLast
 		last = newLast
 		h0, h = h0.next(), h.next()
 		noffset, offset = h0.hash(), h.hash()
 	}
-	last.mask = 1<<noffset | 1<<offset
+	last.mask = 1<<uint(noffset) | 1<<uint(offset)
 	last.children[noffset] = l.node()
 	last.children[offset] = newLeaf(v).node()
 	return result
